requirements.txt
pandas~=1.1.3
numpy
pymongo~=3.9.0
fire~=0.2.1
pillow~=6.2.1
torch_optimizer
randaugment
easydict
tensorboard